+++ src/main/java/LlikelionKNU/KNUfest/domain/comment/service/CommentService.java
@@ -6,8 +6,8 @@ import LlikelionKNU.KNUfest.domain.comment.dto.CommentRequest;
 import java.util.List;
 
 public interface CommentService {
-    List<Comment> getCommentPage(int boothId, int perpage, int page, String order);
-    int postComment(int id, CommentRequest comment);
-    void deleteComment(int commentId, String password);
+    List<Comment> getCommentPage(Long boothId, int perpage, int page, String order);
+    Long postComment(Long id, CommentRequest comment);
+    void deleteComment(Long commentId, String password);
 
 }
